independent field background
