+++ include/picongpu/plugins/shadowgraphy/Shadowgraphy.x.cpp
@@ -17,13 +17,14 @@
  * If not, see <http://www.gnu.org/licenses/>.
  */
 
+// required to get the definition of SIMDIM
+#include "picongpu/simulation_defines.hpp"
 
 #if(SIMDIM == DIM3 && PIC_ENABLE_FFTW3 == 1 && ENABLE_OPENPMD == 1)
 
-#    include "picongpu/plugins/shadowgraphy/Shadowgraphy.hpp"
-
 #    include "picongpu/plugins/PluginRegistry.hpp"
 #    include "picongpu/plugins/multi/multi.hpp"
+#    include "picongpu/plugins/shadowgraphy/Shadowgraphy.hpp"
 
 
 PIC_REGISTER_PLUGIN(picongpu::plugins::multi::Master<picongpu::plugins::shadowgraphy::Shadowgraphy>);
