+++ include/picongpu/simulation/stage/FieldBackground.hpp
@@ -23,15 +23,6 @@
 
 #include "picongpu/simulation_defines.hpp"
 
-#include "picongpu/fields/FieldB.hpp"
-#include "picongpu/fields/FieldE.hpp"
-#include "picongpu/fields/background/cellwiseOperation.hpp"
-
-#include <pmacc/Environment.hpp>
-#include <pmacc/dataManagement/DataConnector.hpp>
-#include <pmacc/math/operation.hpp>
-#include <pmacc/type/Area.hpp>
-
 #include <boost/program_options.hpp>
 
 #include <cstdint>
@@ -47,124 +38,15 @@ namespace picongpu
         {
             namespace detail
             {
-                /* Implementation of background application to the given field
-                 *
-                 * @tparam T_Field field affected, e.g. picongpu::FieldE
-                 * @tparam T_FieldBackground field background to apply, e.g. picongpu::FieldBackgroundE
-                 */
-                template<typename T_Field, typename T_FieldBackground>
-                class ApplyFieldBackground
+                struct IApplyFieldBackground
                 {
-                public:
-                    //! Field affected
-                    using Field = T_Field;
+                    virtual void add(uint32_t const step) = 0;
+                    virtual void subtract(uint32_t const step) = 0;
 
-                    //! Field background to apply
-                    using FieldBackground = T_FieldBackground;
-
-                    /** Create an object to apply the background
-                     *
-                     * @param cellDescription mapping for kernels
-                     * @param useDuplicateField flag to store duplicate of the field
-                     */
-                    ApplyFieldBackground(MappingDesc const cellDescription, bool const useDuplicateField)
-                        : isEnabled(FieldBackground::InfluenceParticlePusher)
-                        , useDuplicateField(useDuplicateField)
-                        , restoreFromDuplicateField(false)
-                        , cellDescription(cellDescription)
-                    {
-                        if(isEnabled && useDuplicateField)
-                        {
-                            // Allocate a duplicate field buffer and copy the values
-                            DataConnector& dc = Environment<>::get().DataConnector();
-                            auto field = dc.get<Field>(Field::getName());
-                            auto const& gridBuffer = field->getGridBuffer();
-                            duplicateBuffer = pmacc::makeDeepCopy(gridBuffer.getDeviceBuffer());
-                        }
-                    }
-
-                    /** Add the field background in the whole local domain
-                     *
-                     * @param step index of time iteration
-                     */
-                    void add(uint32_t const step)
-                    {
-                        if(!isEnabled)
-                            return;
-                        DataConnector& dc = Environment<>::get().DataConnector();
-                        auto& field = *dc.get<Field>(Field::getName());
-                        // Always add to the field, conditionally make a copy of the old values first
-                        if(useDuplicateField)
-                        {
-                            auto& gridBuffer = field.getGridBuffer();
-                            duplicateBuffer->copyFrom(gridBuffer.getDeviceBuffer());
-                            restoreFromDuplicateField = true;
-                        }
-                        apply(step, pmacc::math::operation::Add(), field);
-                    }
-
-                    /** Subtract the field background in the whole local domain
-                     *
-                     * @param step index of time iteration
-                     */
-                    void subtract(uint32_t const step)
-                    {
-                        if(!isEnabled)
-                            return;
-                        DataConnector& dc = Environment<>::get().DataConnector();
-                        auto& field = *dc.get<Field>(Field::getName());
-                        /* Either restore from the pre-made copy or subtract.
-                         * Note that here it is not sufficient to check for useDuplicateField as it
-                         * is not necessarily up-to-date, e.g. right after loading from a checkpoint.
-                         */
-                        if(restoreFromDuplicateField)
-                        {
-                            auto& gridBuffer = field.getGridBuffer();
-                            gridBuffer.getDeviceBuffer().copyFrom(*duplicateBuffer);
-                            restoreFromDuplicateField = false;
-                        }
-                        else
-                            apply(step, pmacc::math::operation::Sub(), field);
-                    }
-
-                private:
-                    //! Is the field background enabled
-                    bool isEnabled;
-
-                    //! Flag to store duplicate of field when the background is enabled
-                    bool useDuplicateField;
-
-                    //! Flag to restore from the duplicate field: true if it is enabled and up-to-date
-                    bool restoreFromDuplicateField;
-
-                    //! Buffer type to store duplicated values
-                    using DeviceBuffer = typename Field::Buffer::DBuffer;
-
-                    //! Buffer to store duplicated values, only used when useDuplicateField is true
-                    std::unique_ptr<DeviceBuffer> duplicateBuffer;
-
-                    //! Mapping for kernels
-                    MappingDesc const cellDescription;
-
-                    /** Apply the given functor to the field background in the whole local domain
-                     *
-                     * @tparam T_Functor functor type compatible to pmacc::math::operation
-                     *
-                     * @param step index of time iteration
-                     * @param functor functor to apply
-                     * @param field field object which data is modified
-                     */
-                    template<typename T_Functor>
-                    void apply(uint32_t const step, T_Functor functor, Field& field)
-                    {
-                        constexpr auto area = CORE + BORDER + GUARD;
-                        using CallBackground = cellwiseOperation::CellwiseOperation<area>;
-                        CallBackground callBackground(cellDescription);
-                        callBackground(&field, functor, FieldBackground(field.getUnit()), step);
-                    }
+                    virtual ~IApplyFieldBackground() = default;
                 };
-            } // namespace detail
 
+            } // namespace detail
             //! Functor for the stage of the PIC loop applying field background
             class FieldBackground
             {
@@ -190,11 +72,7 @@ namespace picongpu
                  *
                  * @param cellDescription mapping for kernels
                  */
-                void init(MappingDesc const cellDescription)
-                {
-                    applyE = std::make_unique<ApplyE>(cellDescription, duplicateFields);
-                    applyB = std::make_unique<ApplyB>(cellDescription, duplicateFields);
-                }
+                void init(MappingDesc const cellDescription);
 
                 /** Add field background to the electromagnetic field
                  *
@@ -203,12 +81,7 @@ namespace picongpu
                  *
                  * @param step index of time iteration
                  */
-                void add(uint32_t const step)
-                {
-                    checkInitialization();
-                    applyE->add(step);
-                    applyB->add(step);
-                }
+                void add(uint32_t const step);
 
                 /** Subtract field background from the electromagnetic field
                  *
@@ -221,12 +94,7 @@ namespace picongpu
                  *
                  * @param step index of time iteration
                  */
-                void subtract(uint32_t const step)
-                {
-                    checkInitialization();
-                    applyE->subtract(step);
-                    applyB->subtract(step);
-                }
+                void subtract(uint32_t const step);
 
             private:
                 //! Check if this class was properly initialized, throws when failed
@@ -236,17 +104,11 @@ namespace picongpu
                         throw std::runtime_error("simulation::stage::FieldBackground used without init() called");
                 }
 
-                //! Implememtation type to apply background to field E
-                using ApplyE = detail::ApplyFieldBackground<FieldE, FieldBackgroundE>;
-
                 //! Object to apply background to field E
-                std::unique_ptr<ApplyE> applyE;
-
-                //! Implememtation type to apply background to field B
-                using ApplyB = detail::ApplyFieldBackground<FieldB, FieldBackgroundB>;
+                std::unique_ptr<detail::IApplyFieldBackground> applyE;
 
                 //! Object to apply background to field B
-                std::unique_ptr<ApplyB> applyB;
+                std::unique_ptr<detail::IApplyFieldBackground> applyB;
 
                 //! Flag to store duplicates fields with enabled backgrounds
                 bool duplicateFields = false;
